// src/interfaces/futureRelease.ts
export interface FutureRelease {
  id:           number;
  title:        string;
  description:  string;
  poster:       string;
  category:     number;
  duration:     string;
  banner:       string;
  synopsis:     string;
  trailer:      string;
  director:     string;
  rating:       string;
  release_date: Date;
}